abstract out validateTemplate
