+++ frontend/packages/system/src/Notifications/Pages/Scheduled/tera.ts
@@ -1,0 +1,25 @@
+import { renderOneOff } from 'tera-web';
+
+const isTemplateError = (err: string) => {
+  return err.startsWith('Template');
+};
+const isParamsError = (err: string) => {
+  return err.startsWith('Parameter');
+};
+
+export const validateTemplate = (template: string) => {
+  try {
+    renderOneOff(template, JSON.stringify({}));
+  } catch (e) {
+    if (typeof e === 'string') {
+      if (isTemplateError(e)) {
+        throw e;
+      }
+      if (isParamsError(e)) {
+        // Missing params are not an error when validating the template
+        return;
+      }
+    }
+    throw 'Unknown error :' + e;
+  }
+};
